fix cavebot spam attack packet bug (again)
currently, the cavebot will spam the same attack packet on the same creature several times while fighting... fix that, so we only send the attack packet if the red square is not around that creature. see https://github.com/blackdtools/Blackd-Proxy-CLASSIC/issues/57  for more info

this bug is probably detectable by cipsoft.